% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_predictor.R
\name{build_predictor}
\alias{build_predictor}
\title{Uses output from select_regions() to build the predictor}
\usage{
build_predictor(inputdata = NULL, phenodata = NULL, phenotype = NULL,
  covariates = NULL, type = NULL, numRegions = NULL)
}
\arguments{
\item{inputdata}{output from select_regions() \code{inputdata}}

\item{phenodata}{data set with phenotype information; samples in rows, 
variables in columns \code{phenodata}}

\item{phenotype}{phenotype of interest \code{phenotype}}

\item{covariates}{Which covariates to include in model \code{covariates}}

\item{type}{The class of the phenotype of interest (numeric, factor) 
\code{type}}

\item{numRegions}{The number of regions per class of variable of interest 
to pull out from each chromosome (default: 10) \code{numRegions}}
}
\value{
An n x m data.frame of coefficient estimates and region indices 
for each of the regions included from select_regions() along with regiondata and indices for trainingProbes
}
\description{
This function takes the output from select_regions() and builds the 
phenotype predictor to be used for phenotype prediction in a new 
data set. This function outputs the fits (coefficient estimates, `coefEsts`)
from the model for the selectied regions as well as the rows selected 
(`trainingProbes`) from the input data.
}
\examples{

library('GenomicRanges')
library('dplyr')

## Make up some some region data
regions <- GRanges(seqnames = 'chr2', IRanges(
     start = c(28971710:28971712, 29555081:29555083, 29754982:29754984),
     end = c(29462417:29462419, 29923338:29923340, 29917714:29917716)))

## make up some expression data for 9 rows and 30 people
data(sysdata, package='phenopredict')   
## includes R object 'cm'
exp= cm[1:length(regions),1:30]

## generate some phenotype information
sex = as.data.frame(rep(c("male","female"),each=15))
age = as.data.frame(sample(1:100,30))
pheno = dplyr::bind_cols(sex,age)
colnames(pheno) <- c("sex","age")

## select regions to be used to build the predictor
inputdata <- select_regions(expression=exp, regiondata=regions,
	phenodata=pheno, phenotype="sex", covariates=NULL,type="factor", numRegions=2)

## build phenotype predictor
predictor<-build_predictor(inputdata=inputdata ,phenodata=pheno, 
	phenotype="sex", covariates=NULL,type="factor", numRegions=2)
}
\keyword{coefficient}
\keyword{estimates}
\keyword{phenotype,}
\keyword{prediction,}
