% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cm}
\alias{cm}
\title{Example epression data used to build the predictor.}
\format{A data frame with 500 rows and 4769 samples (columns):}
\source{
\url{https://bioconductor.org/
packages/release/bioc/html/recount.html}
}
\usage{
cm
}
\description{
A dataset containing expression data for 500 expression regions
on chrX (generated from GTEx data)
}
\keyword{datasets}
