% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cm_new}
\alias{cm_new}
\title{Example expression data to be used for prediction.}
\format{A data frame with 500 rows and 4769 different samples (columns)
than those used in "cm":}
\source{
\url{https://bioconductor.org/packages/release/bioc/html/recount.html}
}
\usage{
cm_new
}
\description{
A dataset containing expression data for 500 expression regions
on chrX (generated from GTEx data)
}
\keyword{datasets}
