% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extracts data for predictions using regions from select_region()}
\usage{
extract_data(expression = NULL, inputdata = NULL, predictordata = NULL)
}
\arguments{
\item{inputdata}{object output from select_regions() \code{inputdata}}

\item{predictordata}{object output from build_predictor \code{predictordata}}

\item{exoression}{expression data set from which to extract coverage data \code{expression}}
}
\value{
comat_test An n x m data.frame of the selected regions from the data set specified by `expression`
}
\description{
This function extracts regions of interest from data set on which you want to make predictions.
}
\examples{
testdata <-extract_data(expression=cm_new, inputdata=inputdata, predictordata=predictor)
}
\keyword{data}
\keyword{phenotype,}
\keyword{prediction,}
\keyword{set}

