% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputdata}
\alias{inputdata}
\title{Output from select_regions}
\format{A list containing three elements
\describe{
  \item{regioninfo}{data.matrix, region information}
  \item{regiondata}{GRanges, genomic ranges region information}
  \item{covmat}{matrix, expression coverage}
}}
\usage{
inputdata
}
\description{
A list including a GRanges object (regiondadta), a expression matrix (cm),
and a dataframe (regioninfo). This is the output of select_regions().
}
\keyword{datasets}

