% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Example phenotype data.}
\format{A data frame with 4769 rows and 109 columns}
\source{
\url{https://bioconductor.org/packages/
release/bioc/html/recount.html}
}
\usage{
pheno
}
\description{
A dataset containing phenotype data for 4769 samples included in "cm"
}
\keyword{datasets}
