% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predictor}
\alias{predictor}
\title{Output from build_predictor}
\format{A list containing two elements
\describe{
  \item{coefEsts}{data.frame, prediction estimates}
  \item{trainingProbes}{vector, probes to be extracted}
}}
\usage{
predictor
}
\description{
A list including a data.frame of coefficient estimates
and a vector of probes used in the model. This is the 
output of build_predictor().
}
\keyword{datasets}

