% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_data.R
\name{predictor_data}
\alias{predictor_data}
\title{Extracts expression info from prediction set (GTEx) using output from select_regions()}
\usage{
predictor_data(phenotype = NULL, chrom = "ALL")
}
\arguments{
\item{phenotype}{phenotype of interest \code{phenotype}}

\item{chrom}{Which chromosomes to include in predictor \code{chrom}}
}
\value{
output Three files: regions_to_subset, coverageMatrixPredictor, and regioninfo
}
\description{
This function takes uses the regions selected from select_regions and extracts the
expression and region information necessary to build the predictor
}
\examples{
predictor_data(phenotype="SMTS",chrom=22)
}
\keyword{phenotype,}
\keyword{prediction,}
\keyword{selection}

