% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predictor_samplesource}
\alias{predictor_samplesource}
\title{Prebuilt sample source predictor}
\format{A list for each predicted phenotype contining 
\describe{
  \item{coefEsts}{data.frame, prediction estimates}
  \item{trainingProbes}{vector, probes to be extracted}
  \item{regiondata}{GRanges, genomic ranges region information}
}}
\usage{
predictor_samplesource
}
\description{
This object includes the output from build_predictor 
for sample source prediction
}
\keyword{datasets}
