% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regiondata}
\alias{regiondata}
\title{Information about regions included}
\format{A GRanges object with 500 ranges}
\usage{
regiondata
}
\description{
A GRanges object with information for the 500 expressed 
regions included in cm and cm_new
}
\keyword{datasets}
