% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_regions.R
\name{select_regions}
\alias{select_regions}
\title{Selects regions from prediction set to be used to build the predctor for a 
phenotype of interest}
\usage{
select_regions(expression = NULL, regiondata = NULL, phenodata = NULL,
  phenotype = NULL, covariates = NULL, type = "factor",
  numRegions = 100)
}
\arguments{
\item{expression}{expression data where regions are in rows and samples are 
in columns \code{expression}}

\item{regiondata}{A GRanges object in which  \code{regiondata}}

\item{phenodata}{phenotype data with samples in rows and corresponding 
phenotype 
information in columns  \code{phenodata}}

\item{phenotype}{phenotype of interest \code{phenotype}}

\item{covariates}{Which covariates to include in model \code{covariates}}

\item{type}{The class of the phenotype of interest (numeric, binary, factor)
\code{type}}

\item{numRegions}{The number of regions per class of variable of interest 
to pull out from each chromosome (default: 100) \code{numRegions}}
}
\value{
res The selected regions, the coverage matrix, and the region info 
to be used for prediction
}
\description{
This function takes phenotype of interest (sex, tissue type, etc.)
input by the user and uses a linear model (accounting for covariates, 
if provided) to select those expressed regions that best predict the 
phenotype of interest. These regions are then used downstream for 
prediction.
}
\keyword{phenotype,}
\keyword{prediction,}
\keyword{selection}

