% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_predictor.R
\name{build_predictor}
\alias{build_predictor}
\title{Uses output from select_regions() to build the predictor}
\usage{
build_predictor(inputdata = NULL, phenodata = NULL, phenotype = NULL,
  covariates = NULL, type = NULL, numRegions = 10)
}
\arguments{
\item{inputdata}{output from select_regions() \code{inputdata}}

\item{phenodata}{data set with phenotype information; samples in rows, 
variables in columns \code{phenodata}}

\item{phenotype}{phenotype of interest \code{phenotype}}

\item{covariates}{Which covariates to include in model \code{covariates}}

\item{type}{The class of the phenotype of interest (numeric, factor) 
\code{type}}

\item{numRegions}{The number of regions per class of variable of interest 
to pull out from each chromosome (default: 10) \code{numRegions}}
}
\value{
An n x m data.frame of coefficient estimates and region indices 
for each of the regions included from select_regions() along with regiondata and indices for trainingProbes
}
\description{
This function takes the output from select_regions() and builds the 
phenotype predictor to be used for phenotype prediction in a new 
data set. This function outputs the fits (coefficient estimates, `coefEsts`)
from the model for the selectied regions as well as the rows selected 
(`trainingProbes`) from the input data.
}
\keyword{coefficient}
\keyword{estimates}
\keyword{phenotype,}
\keyword{prediction,}
