% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extracts data for predictions using regions from select_region()}
\usage{
extract_data(newexpression = NULL, newregiondata = NULL,
  predictordata = NULL)
}
\arguments{
\item{newexpression}{expression data set from which to extract coverage data 
\code{expression}}

\item{newregiondata}{GRanges object containing region info for expression data \code{newregiondata}}

\item{predictordata}{object output from build_predictor \code{predictordata}}
}
\value{
comat_test An n x m data.frame of the selected regions from the 
data set specified by `expression`
}
\description{
This function extracts regions of interest from data set on which you want 
to make predictions.
}
\keyword{data}
\keyword{phenotype,}
\keyword{prediction,}
\keyword{set}

