% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_input.R
\name{merge_input}
\alias{merge_input}
\title{Uses output from select_regions() to merge output before building the 
predictor}
\usage{
merge_input(inputdata_list = NULL)
}
\arguments{
\item{inputdata_list}{list out output objects from select_regions() 
\code{inputdata_list}}
}
\value{
merged inputdata across multiple runs of select_regions()
}
\description{
This function takes the output from multiple executions of select_regions() 
and merges the output before building and running the predictor. Objects
output are the merged output from select_regions but contain the same 
objects.
}
\keyword{expression}
\keyword{merge,}
\keyword{phenotype,}
\keyword{prediction,}
