% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pheno.R
\name{predict_pheno}
\alias{predict_pheno}
\title{Makes predictions for selected phenotype}
\usage{
predict_pheno(expression = NULL, phenodata = NULL, phenotype = NULL,
  predictordata = NULL)
}
\arguments{
\item{expression}{Expression data from test set \code{expression}}

\item{phenodata}{data set with phenotype information; samples in rows, 
variables in columns \code{phenodata}}

\item{phenotype}{phenotype of interest \code{phenotype}}

\item{predictordata}{object output from build_predictor \code{predictordata}}
}
\value{
A vector of predicted phenoyptes
}
\description{
This function outputs predictions for phenotype of 
interest using selected regions upon which predictor 
has been built
}
\keyword{data}
\keyword{phenotype,}
\keyword{prediction,}
\keyword{set}

