% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_predictor.R
\name{test_predictor}
\alias{test_predictor}
\title{Test accuracy of predictor on known phenotypes}
\usage{
test_predictor(inputdata = NULL, phenodata = NULL, phenotype = NULL,
  covariates = NULL, type = "factor", predictordata = NULL)
}
\arguments{
\item{inputdata}{output from select_regions() \code{inputdata}}

\item{phenodata}{data set with phenotype information; samples in rows, variables in columns \code{phenodata}}

\item{phenotype}{phenotype of interest \code{phenotype}}

\item{covariates}{Which covariates to include in model \code{covariates}}

\item{type}{The class of the phenotype of interest (numeric, binary, factor) \code{type}}

\item{predictordata}{object output from build_predictor \code{predictordata}}
}
\value{
res #####
}
\description{
This function takes the expression data input to 
build_predictor() and the coefficient estimates from
build_predictor() for phenotype prediction. The known 
phenotypes are also input for comparison and 
asseessment of predictor accuracy.
}
\examples{
predictions_test <-test_predictor(inputdata=inputdata ,phenodata=pheno, phenotype="Sex", covariates=c("AGE","BMI"),type="factor",predictordata=predictor )
}
\keyword{phenotype,}
\keyword{prediction,}
\keyword{test}

